Refactor the TestMetricAlertsDetectorValidator and MetricAlertComparisonConditionValidatorTest to the new file
